Few things
Added an option to the character to make it randomise every time the player is spawned.
Realised I already HAD a global RNG object and thus was making another for no reason in the Easter Egg message.
Removed descriptions for two removed traps.
Made Gold Gem receives behave differently depending on shops (if Milla's is enabled, then it's a Gold Gem as normal, if not but the Vinyl Shop is, then it gives 100 Crystal Shards instead. If neither is enabled, then it gives a 1UP).