idfk read the desc
Added an option to remove the chest checks (this does require at least one of the shops on to add enough locations).
Fixed not being able to access the Basic Tutorial despite it being on the map (also added back the Guard Tutorial).
Made the server stuff read from the config file (doesn't currently write the last used values to it because I'm lazy and need to do some slight rejigging to do so).