stuff
Made the pause menu on the map show the Time Capsules instead of the Cores.
Made the Gold Gem and Time Capsules pull their descriptions from the game itself (didn't realise they had ones)
Changed the sprites for the Star Cards and Time Capsules to ones from a different texture atlas.
Added a Powerpoint Trap, which drops the game down to 15FPS for 30 seconds.